+++ src/account/operations/balance.rs
@@ -24,8 +24,8 @@ impl AccountHandle {
         let account_addresses = self.addresses().await?;
         let account = self.read().await;
 
-        let network_id = self.client.get_network_id()?;
-        let rent_structure = self.client.get_rent_structure()?;
+        let network_id = self.client.get_network_id().await?;
+        let rent_structure = self.client.get_rent_structure().await?;
 
         let local_time = self.client.get_time_checked()?;