+++ src/account/operations/transaction/high_level/burning_melting/destroy_alias.rs
@@ -79,7 +79,7 @@ impl AccountHandle {
     // governor address from the alias output.
     async fn output_id_and_basic_output_for_alias(&self, alias_id: AliasId) -> crate::Result<(OutputId, Output)> {
         let account = self.read().await;
-        let token_supply = self.client.get_token_supply()?;
+        let token_supply = self.client.get_token_supply().await?;
 
         let (output_id, output_data) = account
             .unspent_outputs()