+++ src/account/operations/transaction/prepare_output.rs
@@ -7,7 +7,7 @@ use iota_client::block::{
     address::Address,
     output::{
         dto::NativeTokenDto,
-        feature::{Feature, MetadataFeature, TagFeature},
+        feature::{Feature, IssuerFeature, MetadataFeature, SenderFeature, TagFeature},
         unlock_condition::{
             AddressUnlockCondition, ExpirationUnlockCondition, StorageDepositReturnUnlockCondition,
             TimelockUnlockCondition, UnlockCondition,
@@ -32,7 +32,7 @@ impl AccountHandle {
         transaction_options: Option<TransactionOptions>,
     ) -> crate::Result<Output> {
         log::debug!("[OUTPUT] prepare_output {options:?}");
-        let token_supply = self.client.get_token_supply()?;
+        let token_supply = self.client.get_token_supply().await?;
 
         if let Some(assets) = &options.assets {
             if let Some(nft_id) = assets.nft_id {
@@ -41,7 +41,7 @@ impl AccountHandle {
                     .await;
             }
         }
-        let rent_structure = self.client.get_rent_structure()?;
+        let rent_structure = self.client.get_rent_structure().await?;
 
         // We start building with minimum storage deposit, so we know the minimum required amount and can later replace
         // it, if needed
@@ -57,13 +57,23 @@ impl AccountHandle {
         }
 
         if let Some(features) = options.features {
+            if let Some(_issuer) = features.issuer {
+                return Err(crate::Error::MissingParameter("nft_id"));
+            }
+
             if let Some(tag) = features.tag {
                 first_output_builder =
                     first_output_builder.add_feature(Feature::Tag(TagFeature::new(tag.as_bytes().to_vec())?));
             }
+
             if let Some(metadata) = features.metadata {
                 first_output_builder = first_output_builder
                     .add_feature(Feature::Metadata(MetadataFeature::new(metadata.as_bytes().to_vec())?));
+            }
+
+            if let Some(sender) = features.sender {
+                first_output_builder = first_output_builder
+                    .add_feature(Feature::Sender(SenderFeature::new(Address::try_from_bech32(sender)?.1)))
             }
         }
 
@@ -112,15 +122,14 @@ impl AccountHandle {
                     second_output_builder = second_output_builder.add_unlock_condition(
                         UnlockCondition::StorageDepositReturn(StorageDepositReturnUnlockCondition::new(
                             remainder_address,
-                            // Return minimum storage deposit + any additional required storage deposit from features
-                            // or unlock conditions
+                            // Return minimum storage deposit
                             min_storage_deposit_return_amount,
                             token_supply,
                         )?),
                     );
                 }
 
-                // Check if the remainding balance wouldn't leave dust behind, which wouldn't allow the creation of this
+                // Check if the remainder balance wouldn't leave dust behind, which wouldn't allow the creation of this
                 // output. If that's the case, this remaining amount will be added to the output, to still allow sending
                 // it.
                 if storage_deposit.use_excess_if_low.unwrap_or_default() {
@@ -185,7 +194,8 @@ impl AccountHandle {
     ) -> crate::Result<Output> {
         log::debug!("[OUTPUT] prepare_nft_output {options:?}");
 
-        let token_supply = self.client.get_token_supply()?;
+        let token_supply = self.client.get_token_supply().await?;
+        let rent_structure = self.client.get_rent_structure().await?;
         let unspent_outputs = self.unspent_outputs(None).await?;
 
         // Find nft output from the inputs
@@ -201,6 +211,8 @@ impl AccountHandle {
             } else {
                 unreachable!("We checked before if it's an nft output")
             }
+        } else if nft_id.is_null() {
+            NftOutputBuilder::new_with_minimum_storage_deposit(rent_structure.clone(), nft_id)?
         } else {
             return Err(crate::Error::NftNotFoundInUnspentOutputs);
         };
@@ -209,10 +221,6 @@ impl AccountHandle {
         first_output_builder = first_output_builder.with_unlock_conditions(vec![UnlockCondition::Address(
             AddressUnlockCondition::new(Address::try_from_bech32(options.recipient_address.clone())?.1),
         )]);
-
-        // from here basically the same as in `prepare_output()`, just with Nft outputs
-
-        let rent_structure = self.client.get_rent_structure()?;
 
         if let Some(assets) = options.assets {
             if let Some(native_tokens) = assets.native_tokens {
@@ -228,6 +236,16 @@ impl AccountHandle {
             if let Some(metadata) = features.metadata {
                 first_output_builder = first_output_builder
                     .add_feature(Feature::Metadata(MetadataFeature::new(metadata.as_bytes().to_vec())?));
+            }
+
+            if let Some(issuer) = features.issuer {
+                first_output_builder = first_output_builder
+                    .add_immutable_feature(Feature::Issuer(IssuerFeature::new(Address::try_from_bech32(issuer)?.1)));
+            }
+
+            if let Some(sender) = features.sender {
+                first_output_builder = first_output_builder
+                    .add_feature(Feature::Sender(SenderFeature::new(Address::try_from_bech32(sender)?.1)))
             }
         }
 
@@ -246,7 +264,9 @@ impl AccountHandle {
             }
         }
 
-        let first_output = first_output_builder.finish(token_supply)?;
+        let first_output = first_output_builder
+            .with_minimum_storage_deposit(rent_structure.clone())
+            .finish(token_supply)?;
 
         let mut second_output_builder = NftOutputBuilder::from(&first_output);
 
@@ -276,8 +296,7 @@ impl AccountHandle {
                     second_output_builder = second_output_builder.add_unlock_condition(
                         UnlockCondition::StorageDepositReturn(StorageDepositReturnUnlockCondition::new(
                             remainder_address,
-                            // Return minimum storage deposit + any additional required storage deposit from features
-                            // or unlock conditions
+                            // Return minimum storage deposit
                             min_storage_deposit_return_amount,
                             token_supply,
                         )?),
@@ -393,6 +412,8 @@ pub struct Assets {
 pub struct Features {
     pub tag: Option<String>,
     pub metadata: Option<String>,
+    pub issuer: Option<String>,
+    pub sender: Option<String>,
 }
 
 #[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]